Day 8/#100DaysOfCode
The profile lookup project was a little frustrating, because I don't understand why that one line of code had to be outside the for loop. Although as I was typing that I think I got a brainwave and understood. It was probably exiting the for loop with a return statement too early and not getting to access the entire array. Okay, I get it.